Add Splash Screen to Windows
fix #1160
