+++ Src/MoneyFox.Windows/App.xaml.cs
@@ -65,64 +65,21 @@ namespace MoneyFox.Windows
         ///     will be used such as when the application is launched to open a specific file.
         /// </summary>
         /// <param name="e">Details about the launch request and process.</param>
-        protected override async void OnLaunched(LaunchActivatedEventArgs e)
+        protected override void OnLaunched(LaunchActivatedEventArgs e)
         {
             var shell = Window.Current.Content as AppShell;
 
-            // Do not repeat app initialization when the Window already has content,
-            // just ensure that the window is active
-            if (shell == null)
+            if (e.PreviousExecutionState != ApplicationExecutionState.Running)
             {
-                //TODO: check if this can be removed with the neutral language set.
-                // Create a AppShell to act as the navigation context and navigate to the first page
-                shell = new AppShell {Language = ApplicationLanguages.Languages[0]};
-
-                shell.MyAppFrame.NavigationFailed += OnNavigationFailed;
+                bool loadState = (e.PreviousExecutionState == ApplicationExecutionState.Terminated);
+                var extendedSplash = new ExtendedSplashScreen(e.SplashScreen, loadState);
+                Window.Current.Content = extendedSplash;
             }
-
-            // Place our app shell in the current Window
-            Window.Current.Content = shell;
-            ApplicationLanguages.PrimaryLanguageOverride = GlobalizationPreferences.Languages[0];
-
-            if (shell.MyAppFrame.Content == null)
-            {
-                // When the navigation stack isn't restored, navigate to the first page
-                // suppressing the initial entrance animation.
-                var setup = new Setup(shell.MyAppFrame);
-                setup.Initialize();
-
-                var start = Mvx.Resolve<IMvxAppStart>();
-                start.Start();
-
-                shell.ViewModel = Mvx.Resolve<MenuViewModel>();
-            }
-
-            new TileHelper().DoNavigation(string.IsNullOrEmpty(e.Arguments)
-                ? e.TileId
-                : e.Arguments);
 
             OverrideTitleBarColor();
-
-            //If Jump Lists are supported, adds them§a
-            if (ApiInformation.IsTypePresent("Windows.UI.StartScreen.JumpList"))
-            {
-                SetJumplist();
-            }
-
-            CallRateReminder();
-
+            
             // Ensure the current window is active
             Window.Current.Activate();
-        }
-
-        private async void CallRateReminder()
-        {
-            RatePopup.RateButtonText = Strings.YesLabel;
-            RatePopup.CancelButtonText = Strings.NotNowLabel;
-            RatePopup.Title = Strings.RateReminderTitle;
-            RatePopup.Content = Strings.RateReminderText;
-
-            await RatePopup.CheckRateReminderAsync();
         }
 
         private void OverrideTitleBarColor()
@@ -151,40 +108,6 @@ namespace MoneyFox.Windows
                 titleBar.ForegroundColor = Colors.White;
                 titleBar.ButtonForegroundColor = Colors.White;
             }
-        }
-
-        private async void SetJumplist()
-        {
-            var jumpList = await JumpList.LoadCurrentAsync();
-            jumpList.Items.Clear();
-            jumpList.SystemGroupKind = JumpListSystemGroupKind.None;
-
-            var listItemAddIncome = JumpListItem.CreateWithArguments(Constants.ADD_INCOME_TILE_ID,
-                Strings.AddIncomeLabel);
-            listItemAddIncome.Logo = new Uri("ms-appx:///Assets/IncomeTileIcon.png");
-            jumpList.Items.Add(listItemAddIncome);
-
-            var listItemAddSpending = JumpListItem.CreateWithArguments(Constants.ADD_EXPENSE_TILE_ID,
-                Strings.AddExpenseLabel);
-            listItemAddSpending.Logo = new Uri("ms-appx:///Assets/SpendingTileIcon.png");
-            jumpList.Items.Add(listItemAddSpending);
-
-            var listItemAddTransfer = JumpListItem.CreateWithArguments(Constants.ADD_TRANSFER_TILE_ID,
-                Strings.AddTransferLabel);
-            listItemAddTransfer.Logo = new Uri("ms-appx:///Assets/TransferTileIcon.png");
-            jumpList.Items.Add(listItemAddTransfer);
-
-            await jumpList.SaveAsync();
-        }
-
-        /// <summary>
-        ///     Invoked when Navigation to a certain page fails
-        /// </summary>
-        /// <param name="sender">The Frame which failed navigation</param>
-        /// <param name="e">Details about the navigation failure</param>
-        private void OnNavigationFailed(object sender, NavigationFailedEventArgs e)
-        {
-            throw new Exception("Failed to load Page " + e.SourcePageType.FullName);
         }
 
         /// <summary>
