+++ Src/MoneyFox.Windows/Style/ColorsDark.xaml
@@ -6,6 +6,8 @@
     <Color x:Key="ColorPrimary">#263873</Color>
     <Color x:Key="ColorPrimaryDark">#1a264c</Color>
 
+    <Color x:Key="ColorSplashScreenBackground">#FFEFEADD</Color>
+
     <Color x:Key="ColorAppBar">#111632</Color>
 
     <Color x:Key="AppForegroundColorPrimary">White</Color>
