+++ Src/MoneyFox.Windows/Style/ColorsLight.xaml
@@ -1,11 +1,12 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
-    <!--Set Accent color manually-->
+    <!--  Set Accent color manually  -->
     <Color x:Key="SystemAccentColor">#314a9b</Color>
 
     <Color x:Key="ColorPrimary">#263873</Color>
     <Color x:Key="ColorPrimaryDark">#1a264c</Color>
+
+    <Color x:Key="ColorSplashScreenBackground">#FFEFEADD</Color>
 
     <Color x:Key="ColorAppBar">#111632</Color>
