+++ Src/MoneyFox.Windows/Views/ExtendedSplashScreen.xaml
@@ -1,0 +1,20 @@
+﻿<Page x:Class="MoneyFox.Windows.Views.ExtendedSplashScreen"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+      mc:Ignorable="d">
+
+    <Grid Background="{StaticResource SplashScreenBackgroundBrush}">
+        <StackPanel VerticalAlignment="Center">
+            <Image Width="150"
+                   Height="150"
+                   Source="../Assets/Square310x310Logo.scale-400.png" />
+
+            <ProgressRing Width="50"
+                          Height="50"
+                          Margin="0,20,0,0"
+                          IsActive="True" />
+        </StackPanel>
+    </Grid>
+</Page>
